Move start_orb_run to orb.cc (wheals)
