+++ crawl-ref/source/orb.h
@@ -4,4 +4,6 @@
 void orb_pickup_noise(const coord_def& where, int loudness = 30,
                       const char* msg = nullptr, const char* msg2 = nullptr);
 
+void start_orb_run(game_chapter new_chapter, const char* message);
+
 #endif
